Update script.js
Continued working on displaying the bookmarks to resultsText. We fixed (hopefully) an issue that occurred when attempting to add functionality to bold a part of the bookmark titles that contained the user's input, searchText, in which the bookmark titles were not displaying correctly; it turned out that this was a case sensitivity issue (I think), since when we tried locating the indices of where searchText began and ended in a bookmark title, we forgot to convert both searchText and the title to lowercase.

Began adding docstrings and more comments!