Update style.css
Revised the styling of the search bar (and maybe the HTML and body???)

We had an issue where, upon changing themes, the new theme background would only cover most of the background, except for the area above the search bar. After doing a bit of debugging, it seems that what was happening was the <body> of the extension was not reaching the top of the background, only seeming to reach up to the top of the search bar. I found that getting rid of the "display: block" property in the search bar styling helped solve this issue, which is what we have for now.