Update script.js
Made a lot more progress on the theme-switching functionality! We are currently, to my knowledge, able to click on the cloud and moon icons on the bottom-right of the extension window, and the backgrounds and texts of the extension can change. I also started implementing functionality to keep the theme we have at a given point, as sometimes, if we switch to the night theme before entering anything into the search bar, when we then search for a bookmark, the links will still retain the color of the sky theme, despite the extension being in the night theme at that moment.