Update background.js
I'm working on adding functionality so that the extension can "remember" the most recent theme it was in before the user closes the extension, so it can then display that theme again when the user opens the extension again. Hence, we will be using chrome's local storage to store this information.

We're finally using background.js wooo!!!

Note: chrome's local storage will store the sky theme by default upon installation (and I think reload??? I could be wrong) of the extension.