Update script.js
We encountered an issue where the placeholder text of the search bar was not changing colors according to the theme the extension was in. We were eventually able to find a solution using variables in CSS, and for the time being, we have added some quick variable changes in script.js to test out this solution!

The solution seems to work! We just need to clean the code up a little