Update script.js
Added docstrings for the hoverPencil, unhoverPencil, and stationaryTheme functions

Changed the element colors a little for the stationary theme

Added a few console logs to the DevTools console for debugging (if needed in the future)

Currently, I'd say the extension is in a decent place! I might work on adding more functionality in the future, however I think I will leave it as it is for now!